refactor and add gitignore
