+++ examples/flask_gevent/flask_app.py
@@ -1,27 +1,37 @@
-from flask import Flask, make_response
-from flask_sockets import Sockets
-from graphql_ws.server import GeventSubscriptionServer
 import json
-from template import render_graphiql
+
 import graphene
-import gevent
+from flask import Flask, make_response
 from flask_graphql import GraphQLView
-import asyncio
+from flask_sockets import Sockets
 from rx import Observable
+
+from graphql_ws import GeventSubscriptionServer
+from template import render_graphiql
 
 
 class Query(graphene.ObjectType):
     base = graphene.String()
 
 
+class RandomType(graphene.ObjectType):
+    seconds = graphene.Int()
+    random_int = graphene.Int()
+
+
 class Subscription(graphene.ObjectType):
 
-    count_seconds = graphene.String()
+    count_seconds = graphene.Int()
+
+    random_int = graphene.Field(RandomType)
 
 
     def resolve_count_seconds(root, info):
         return Observable.interval(1000).map(lambda i: "{0}".format(i))
 
+    def resolve_random_int(root, info):
+        import random
+        return Observable.interval(1000).map(lambda i: RandomType(seconds=i, random_int=random.randint(0, 500)))
 
 schema = graphene.Schema(query=Query, subscription=Subscription)
