+++ examples/flask_gevent/template.py
@@ -119,7 +119,7 @@ def render_graphiql():
 </html>''').substitute(
         GRAPHIQL_VERSION='0.10.2',
         SUBSCRIPTIONS_TRANSPORT_VERSION='0.7.0',
-        subscriptionsEndpoint='ws://localhost:8000/subscriptions',
+        subscriptionsEndpoint='ws://localhost:5000/subscriptions',
         # subscriptionsEndpoint='ws://localhost:5000/',
         endpointURL='/graphql',
     )
