+++ graphql_ws/__init__.py
@@ -9,6 +9,6 @@ __version__ = '0.1.0'
 
 from .observable_aiter import setup_observable_extension
 from .server import WebSocketSubscriptionServer
-
+from .gevent_server import GeventSubscriptionServer
 
 setup_observable_extension()
