+++ graphql_ws/gevent_server.py
@@ -1,0 +1,120 @@
+import json
+
+from graphql import format_error, graphql
+from graphql.execution import ExecutionResult
+from graphql.execution.executors.sync import SyncExecutor
+from rx import Observer
+from promise import is_thenable, Promise
+from .server import BaseWebSocketSubscriptionServer, ConnectionContext, ConnectionClosedException
+from .constants import *
+
+
+class GEventConnectionContext(ConnectionContext):
+
+    def receive(self):
+        msg = self.ws.receive()
+        return msg
+
+    def send(self, data):
+        if self.closed:
+            return
+        self.ws.send(data)
+
+    @property
+    def closed(self):
+        return self.ws.closed
+
+    def close(self, code):
+        self.ws.close(code)
+
+class GeventSubscriptionServer(BaseWebSocketSubscriptionServer):
+
+    def get_graphql_params(self, *args, **kwargs):
+        params = super(GeventSubscriptionServer, self).get_graphql_params(*args, **kwargs)
+        return dict(params, executor=SyncExecutor())
+
+    def handle(self, ws):
+        connection_context = GEventConnectionContext(ws)
+        self.on_open(connection_context)
+        while True:
+            try:
+                if connection_context.closed:
+                    raise ConnectionClosedException()
+                message = connection_context.receive()
+            except ConnectionClosedException:
+                self.on_close(connection_context)
+                return
+            self.on_message(connection_context, message)
+
+    def on_message(self, connection_context, message):
+        try:
+            parsed_message = json.loads(message)
+            assert isinstance(
+                parsed_message, dict), "Payload must be an object."
+        except Exception as e:
+            self.send_error(connection_context, None, e)
+            return
+
+        self.process_message(connection_context, parsed_message)
+
+    def on_open(self, connection_context):
+        pass
+
+    def on_connect(self, connection_context, payload):
+        pass
+
+    def on_close(self, connection_context):
+        remove_operations = list(connection_context.operations.keys())
+        for op_id in remove_operations:
+            self.unsubscribe(connection_context, op_id)
+
+    def on_connection_init(self, connection_context, op_id, payload):
+        try:
+            self.on_connect(connection_context, payload)
+            self.send_message(connection_context, op_type=GQL_CONNECTION_ACK)
+
+        except Exception as e:
+            self.send_error(connection_context, op_id, e, GQL_CONNECTION_ERROR)
+            connection_context.close(1011)
+
+    def on_connection_terminate(self, connection_context, op_id):
+        connection_context.close(1011)
+
+    
+    def on_start(self, connection_context, op_id, params):
+        try:
+            execution_result = graphql(
+                self.schema, **params, allow_subscriptions=True
+            )
+            execution_result.subscribe(SubscriptionObserver(
+                    connection_context,
+                    op_id,
+                    self.send_execution_result,
+                    self.send_error,
+                    self.on_close
+                )
+            )
+        except Exception as e:
+            self.send_error(connection_context, op_id, str(e))
+
+    def on_stop(self, connection_context, op_id):
+        self.unsubscribe(connection_context, op_id)
+
+
+class SubscriptionObserver(Observer):
+
+    def __init__(self, connection_context, op_id, send_execution_result, send_error, on_close):
+        self.connection_context = connection_context
+        self.op_id = op_id
+        self.send_execution_result = send_execution_result
+        self.send_error = send_error
+        self.on_close = on_close
+
+    def on_next(self, value):
+        self.send_execution_result(self.connection_context, self.op_id, value)
+
+    def on_completed(self):
+        self.on_close(self.connection_context)
+    
+    def on_error(self, error):
+        self.send_error(self.connection_context, self.op_id, error)
